[Ref] Stream Export: Avoid theoretical 64->32 truncation.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18762 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Mod2wave.cpp
@@ -1342,7 +1342,7 @@ void CDoWaveConvert::Run()
 				break;
 			}
 			const std::streampos newPos = fileStream.tellp();
-			bytesWritten += static_cast<std::size_t>(newPos - oldPos);
+			bytesWritten += static_cast<uint64>(newPos - oldPos);
 
 			auto currentTime = timeGetTime();
 			if((currentTime - prevTime) >= 16)
